+++ src/converse.js
@@ -35,7 +35,10 @@ import "./plugins/singleton.js";
 import "./components/converse.js";
 import "../sass/converse.scss";
 
-import { converse } from "@converse/headless/core";
+import { _converse, converse } from "@converse/headless/core";
+import { CustomElement } from 'components/element';
+
+_converse.CustomElement = CustomElement;
 
 const WHITELISTED_PLUGINS = [
     'converse-bookmark-views',