+++ src/headless/core.js
@@ -15,7 +15,6 @@ import sizzle from 'sizzle';
 import u from '@converse/headless/utils/core';
 import { Collection } from "@converse/skeletor/src/collection";
 import { Connection, MockConnection } from '@converse/headless/shared/connection.js';
-import { CustomElement } from '../components/element';
 import { Events } from '@converse/skeletor/src/events.js';
 import { Model } from '@converse/skeletor/src/model.js';
 import { Router } from '@converse/skeletor/src/router.js';
@@ -1650,7 +1649,6 @@ Object.assign(converse, {
         $pres,
         'utils': u,
         Collection,
-        CustomElement,
         Model,
         Promise,
         Strophe,